User Select Start Trip
The user will select a team from the drop down box (QComboBox) and it will create the shortest path from the selected stadium to all the other stadiums in the graph using Dijkstra's algo. It will show the total distance of the trip and it will allow the user to see the cart in order to purchase items.